Fix remaining type errors in hover, minimap, mem
Completes the type correctness refactoring by addressing remaining
conversion warnings in GUI and platform-specific code.

Changes:
- hover.c: Changed last_tick from int to uint32_t to match tick type.
  Fixed col variable to unsigned short to match render_text() signature
  (16-bit RGB555 color format). Removed unnecessary sign assumptions in
  experience calculations.

- minimap.c: Fixed set_pix() to use unsigned char for map pixel values
  (matches _mmap array type). Changed map_compare() and map_merge() to
  use unsigned char* for correct pointer types. Introduced local mn
  variable to reduce repeated array index calculations and improve
  clarity. Fixed DIST comparisons with explicit casts where signed
  arithmetic is needed.

- memory_linux.c: Added explicit casts for sysconf() return values to
  prevent sign conversion warnings when multiplying long values.

- astonia.h: Added missing function prototypes for get_total_system_memory,
  get_memory_usage, and client_version to eliminate -Wmissing-prototypes
  warnings.

- version.c: Added astonia.h include for prototype.

+++ src/game/memory_linux.c
@@ -1,14 +1,16 @@
 #include <sys/resource.h>
 #include <unistd.h>
 
+#include "astonia.h"
+
 unsigned long long get_total_system_memory(void)
 {
-	return sysconf(_SC_PHYS_PAGES) * sysconf(_SC_PAGE_SIZE);
+	return (unsigned long long)sysconf(_SC_PHYS_PAGES) * (unsigned long long)sysconf(_SC_PAGE_SIZE);
 }
 
 size_t get_memory_usage(void)
 {
 	struct rusage r_usage;
 	getrusage(RUSAGE_SELF, &r_usage);
-	return r_usage.ru_maxrss * 1024;
+	return (size_t)r_usage.ru_maxrss * 1024;
 }
